Split TdsEnums.cs Add references where necessary
