+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Connection/ConnectionCapabilities.cs
@@ -5,6 +5,7 @@
 using Microsoft.Data.SqlClient.Internal;
 using System;
 using System.Text;
+using Microsoft.Data.SqlClient.Parser;
 
 #nullable enable
