+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ManagedSni/LocalDB.netcore.cs
@@ -12,6 +12,7 @@ using Interop.Windows.Kernel32;
 using Microsoft.Win32;
 using Microsoft.Data.SqlClient.Internal;
 using Microsoft.Data.Common;
+using Microsoft.Data.SqlClient.Parser;
 
 namespace Microsoft.Data.SqlClient.ManagedSni
 {
