+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlBuffer.cs
@@ -7,6 +7,7 @@ using System.Data.SqlTypes;
 using System.Diagnostics;
 using System.Globalization;
 using System.Runtime.InteropServices;
+using Microsoft.Data.SqlClient.Parser;
 using Microsoft.Data.SqlTypes;
 
 namespace Microsoft.Data.SqlClient
@@ -918,7 +919,7 @@ namespace Microsoft.Data.SqlClient
                     {
                         return SqlMoney.Null;
                     }
-                    
+
                     #if NET
                     return SqlMoney.FromTdsValue(_value._int64);
                     #else
@@ -1270,7 +1271,7 @@ namespace Microsoft.Data.SqlClient
                 }
             }
 
-            return null; // need to return the value as an object of some CLS type            
+            return null; // need to return the value as an object of some CLS type
         }
 
         internal static SqlBuffer[] CreateBufferArray(int length)
@@ -1478,7 +1479,7 @@ namespace Microsoft.Data.SqlClient
             }
             timeInfo._ticks = tickUnits * TdsEnums.TICKS_FROM_SCALE[scale];
 
-            // Once the deserialization has been completed using the value scale, we need to set the actual denormalized scale, 
+            // Once the deserialization has been completed using the value scale, we need to set the actual denormalized scale,
             // coming from the data type, on the original result, so that it has the proper scale setting.
             // This only applies for values that got serialized/deserialized for encryption. Otherwise, both scales should be equal.
             timeInfo._scale = denormalizedScale;
@@ -1499,16 +1500,16 @@ namespace Microsoft.Data.SqlClient
         }
         // [Field]As<T> method explanation:
         // these methods are used to bridge generic to non-generic access to value type fields on the storage struct
-        // where typeof(T) == typeof(field) 
-        //   1) RyuJIT will recognize the pattern of (T)(object)T as being redundant and eliminate 
-        //   the T and object casts leaving T, so while this looks like it will put every value type instance in a box the 
+        // where typeof(T) == typeof(field)
+        //   1) RyuJIT will recognize the pattern of (T)(object)T as being redundant and eliminate
+        //   the T and object casts leaving T, so while this looks like it will put every value type instance in a box the
         //   generated assembly will be short and direct
         //   2) another jit may not recognize the pattern and should emit the code as seen. this will box and then unbox the
         //   value type which is no worse than the mechanism that this code replaces
         // where typeof(T) != typeof(field)
         //   the jit will emit all the cast operations as written. this will put the value into a box and then attempt to
-        //   cast it, because it is an object no conversions are used and this will generate the desired InvalidCastException       
-        //   for example users cannot widen a short to an int preserving external expectations 
+        //   cast it, because it is an object no conversions are used and this will generate the desired InvalidCastException
+        //   for example users cannot widen a short to an int preserving external expectations
 
         internal T ByteAs<T>()
         {
