+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlConnectionStringBuilder.cs
@@ -16,6 +16,7 @@ using System.Globalization;
 using System.Reflection;
 using Microsoft.Data.Common;
 using Microsoft.Data.Common.ConnectionString;
+using Microsoft.Data.SqlClient.Parser;
 
 namespace Microsoft.Data.SqlClient
 {
