+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlEnums.cs
@@ -16,6 +16,7 @@ using System.Globalization;
 using System.IO;
 using System.Xml;
 using Microsoft.Data.Common;
+using Microsoft.Data.SqlClient.Parser;
 using Microsoft.Data.SqlTypes;
 using Microsoft.SqlServer.Server;
 
@@ -120,7 +121,7 @@ namespace Microsoft.Data.SqlClient
             type == SqlDbType.Char || type == SqlDbType.VarChar || type == SqlDbType.Text;
 
         // is this type size expressed as count of characters or bytes?
-        private static bool _IsSizeInCharacters(SqlDbType type) => 
+        private static bool _IsSizeInCharacters(SqlDbType type) =>
             type == SqlDbType.NChar ||
             type == SqlDbType.NVarChar ||
             type == SqlDbType.Xml ||
@@ -317,7 +318,7 @@ namespace Microsoft.Data.SqlClient
         //
         internal static MetaType GetMetaTypeFromType(Type dataType) => GetMetaTypeFromValue(dataType, null, false, true);
 
-        internal static MetaType GetMetaTypeFromValue(object value, bool streamAllowed = true) => 
+        internal static MetaType GetMetaTypeFromValue(object value, bool streamAllowed = true) =>
             GetMetaTypeFromValue(value.GetType(), value, true, streamAllowed);
 
         private static MetaType GetMetaTypeFromValue(Type dataType, object value, bool inferLen, bool streamAllowed)
@@ -694,7 +695,7 @@ namespace Microsoft.Data.SqlClient
             Debug.Assert(attributes.Length > 0, failedAssertMessage);
         }
 
-        // devnote: This method should not be used with SqlDbType.Date and SqlDbType.DateTime2. 
+        // devnote: This method should not be used with SqlDbType.Date and SqlDbType.DateTime2.
         //          With these types the values should be used directly as CLR types instead of being converted to a SqlValue
         internal static object GetSqlValueFromComVariant(object comVal)
         {
