+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/TransactionRequest.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Data.SqlClient.Parser;
+
 namespace Microsoft.Data.SqlClient;
 
 /// <summary>
