+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlTypes/SqlVector.cs
@@ -10,6 +10,7 @@ using System.Runtime.InteropServices;
 using System.Text.Json;
 using Microsoft.Data.Common;
 using Microsoft.Data.SqlClient;
+using Microsoft.Data.SqlClient.Parser;
 
 #nullable enable
